No Action Masking Baseline.
Single thread support and heterogeneous scen added
Created CybORG_DummyVecEnv to support single thread training. Should be useful for debugging and testing.
Added a new scenario file for heterogeneous scenarios. The agents are heterogeneous in terms of actions with one having access to the Block action while the other has access to Sleep.

This should be the final commit for this branch. Can be used for training and testing the CommFormer agents in a no action masking scenario.